Update configuration to demonstrate problem with injected properties
The jib plugin produces a correct image now, even if the warName property of
the maven-war-plugin is used. Even using a maven property for the warName
property works for the jib plugin works as long as the maven property is
defined in the pom itself.
Creating the property using the properties-maven-plugin however causese the
build to fail as the property reference seems not to be resolved correctly.